+++ examples/resources/nxos_bgp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_bgp.example "sys/bgp"