+++ examples/resources/nxos_bgp/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_bgp" "example" {
+  admin_state = "enabled"
+}