+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.3 (unreleased)
 
 - Add nxos_feature_hmm resource and data source
+- Add nxos_bgp resource and data source
 
 ## 0.3.2